Improve build version stuff

+++ paper/src/main/java/rocks/gravili/notquests/paper/structs/actions/NumberAction.java
@@ -113,7 +113,7 @@ public class NumberAction extends Action {
                     .meta(CommandMeta.DESCRIPTION, "Creates a new Number action")
                     .handler((context) -> {
 
-                        final int amount = context.get("amount");
+                        final long amount = context.get("amount");
 
                         final String mathOperator = context.get("operator");
